make solver a standalone module
